The JSON request /admin/cmspagemove is corrected and improved: - the page is correctly move according to the provided parameters - a new parameter (mode) was introduced; its possible values are: "append", "before" and "after"
Examples:

http://localhost:8080/roda/admin/cmspagemove?group=&id=15&mode=append
http://localhost:8080/roda/admin/cmspagemove?group=16&id=15&mode=after
http://localhost:8080/roda/admin/cmspagemove?group=8&id=15&mode=append
http://localhost:8080/roda/admin/cmspagemove?group=10&id=15&mode=before

For the moment, "before" and "after" have the same behavior: the page is
moved at the same level as the page specified in the parameter "group"
(i.e. under the same parent as the "group" page). If the value "append"
is specified, then the page is moved under the "group" page (i.e. the
"group" page becomes its parent).